Add files via upload
Stress and depression are significant issues that threaten mental health. People often share their emotions on social media platforms like Instagram through posts and stories. In this project, we will utilize machine learning algorithms to predict the sentiment of the messages shared by users. Our aim is to create positive and negative word clouds based on the sentiment analysis results. Initially, the data was in a text file format, which we converted to CSV for easier processing and analysis. By visualizing the sentiments, we hope to gain insights into the emotional expressions of individuals on social media.